NOTE: (old lines) / (new lines)
Major edits:
1. Made the Rag(x)df with the correct column names from the start instead of changing/adding them later on
2. Changed the base R histogram to a ggplot
3. Made a function and applied lapply to shorten the code when making the classifier

Minor edits:
1. RSQLite package
2. Rag to RAG

--Details:
line 21-22 (new): added RSQLite package, which is required by DECIPHER

lines 39/41, 45/47, 51/53, 68/70, 73/75, 85/87, 150/131, 233/214: replacing Rag with RAG in search and comments. I double check-ed to make sure the search results were the same.

lines 90-93 / 92-95 AND 120-122 / 116-118: Changing the column titles so that they will not have to be renamed or added later on.

Because of this above edit, I removed lines 98-103 AND 127-141.

Lines 155-167 / 136-148: Replacing the base R histogram with a ggplot version.

Lines 229-233 (new): Making a function (named classifiertrial) to shorten the randomForest() code that is repeated throughout since the same parameters are used each time.

Lines 237-240 (new): making a list object to pass each trial of predictors that you did through randomForest. I also added the trinucleotides to see the whole picture (although we already know it will be perfect).

Because of the above edits, code that uses the full randomForest() is removed and replaced with lapply (new lines 242-243)

lines 266 / 252: replacing the results now that they are held in the list, results.

lines 284-293 / 270-274: again, replacing the old code to to new code that uses lapply with our new classifiertrial function.

I made sure to check that the new function and the lapply onto the list had the same results in the original code using all.equal() and checking the results manually.